Revert "Bundled Vote fixes" ( leaving abi fixes ) (#656)
* Revert "Bundled Vote fixes"

* Update fio.address.cpp

* Update fio.address.cpp

* Update fio.fee.abi

* Update fio.address.cpp

+++ fio.contracts/contracts/fio.address/fio.address.cpp
@@ -132,7 +132,7 @@ namespace fioio {
         }
 
         inline uint64_t getBundledAmount() {
-            uint128_t totalcount = 0;
+            int totalcount = 0;
             uint64_t returnvalue = 0;
 
             if (bundlevoters.end() == bundlevoters.begin()) {
